+++ test/Konsola.Tests/Konsola.Tests.csproj
@@ -39,12 +39,12 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Metadata\MetadataTests.cs" />
-    <Compile Include="Parsing\CommandLineParserTests.cs" />
-    <Compile Include="Parsing\ConstraintTests.cs" />
+    <Compile Include="Parser\CommandLineParserTests.cs" />
+    <Compile Include="Parser\ConstraintTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Parsing\UtilTests.cs" />
-    <Compile Include="Parsing\IConsole.Test.cs" />
-    <Compile Include="Parsing\_Contexts.cs" />
+    <Compile Include="Parser\UtilTests.cs" />
+    <Compile Include="Parser\IConsole.Test.cs" />
+    <Compile Include="Parser\_Contexts.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Konsola\Konsola.csproj">
@@ -57,6 +57,10 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <Reference Include="Moq, Version=4.2.1502.911, Culture=neutral, PublicKeyToken=69f491c39445e920, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Moq.4.2.1502.0911\lib\net40\Moq.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="xunit.abstractions, Version=2.0.0.0, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c, processorArchitecture=MSIL">
       <HintPath>..\..\packages\xunit.abstractions.2.0.0\lib\net35\xunit.abstractions.dll</HintPath>
