+++ test/Konsola.Tests/Parser/ConstraintTests.cs
