+++ test/Konsola.Tests/Parser/IConsole.Test.cs
