+++ test/Konsola.Tests/Parser/_Contexts.cs
